Break out and rename translate_gates_to_pulse to lowering.py as lower_gates. This will prevent difficult merge conflicts later when circuit delays are added
