+++ qiskit/scheduler/methods/basic.py
@@ -13,25 +13,15 @@
 """
 The most straightforward scheduling methods: scheduling **as early** or **as late** as possible.
 """
-
-from collections import defaultdict, namedtuple
+from collections import defaultdict
 from typing import List
 
-from qiskit.circuit.measure import Measure
 from qiskit.circuit.quantumcircuit import QuantumCircuit
-from qiskit.exceptions import QiskitError
 from qiskit.circuit.barrier import Barrier
-from qiskit.pulse.channels import AcquireChannel
-from qiskit.pulse.exceptions import PulseError
-from qiskit.pulse.macros import measure
 from qiskit.pulse.schedule import Schedule
 
 from qiskit.scheduler.config import ScheduleConfig
-
-
-CircuitPulseDef = namedtuple('CircuitPulseDef', [
-    'schedule',  # The schedule which implements the quantum circuit command
-    'qubits'])   # The labels of the qubits involved in the command according to the circuit
+from qiskit.scheduler.lowering import lower_gates
 
 
 def as_soon_as_possible(circuit: QuantumCircuit,
@@ -61,7 +51,7 @@ def as_soon_as_possible(circuit: QuantumCircuit,
             qubit_time_available[q] = time
 
     start_times = []
-    circ_pulse_defs = translate_gates_to_pulse_defs(circuit, schedule_config)
+    circ_pulse_defs = lower_gates(circuit, schedule_config)
     for circ_pulse_def in circ_pulse_defs:
         start_time = max(qubit_time_available[q] for q in circ_pulse_def.qubits)
         stop_time = start_time
@@ -106,7 +96,7 @@ def as_late_as_possible(circuit: QuantumCircuit,
             qubit_time_available[q] = time
 
     rev_stop_times = []
-    circ_pulse_defs = translate_gates_to_pulse_defs(circuit, schedule_config)
+    circ_pulse_defs = lower_gates(circuit, schedule_config)
     for circ_pulse_def in reversed(circ_pulse_defs):
         start_time = max(qubit_time_available[q] for q in circ_pulse_def.qubits)
         stop_time = start_time
@@ -121,69 +111,3 @@ def as_late_as_possible(circuit: QuantumCircuit,
     timed_schedules = [(time, cpd.schedule) for time, cpd in zip(start_times, circ_pulse_defs)
                        if not isinstance(cpd.schedule, Barrier)]
     return Schedule(*timed_schedules, name=circuit.name)
-
-
-def translate_gates_to_pulse_defs(circuit: QuantumCircuit,
-                                  schedule_config: ScheduleConfig) -> List[CircuitPulseDef]:
-    """
-    Return a list of Schedules and the qubits they operate on, for each element encountered in th
-    input circuit.
-
-    Without concern for the final schedule, extract and return a list of Schedules and the qubits
-    they operate on, for each element encountered in the input circuit. Measures are grouped when
-    possible, so ``qc.measure(q0, c0)`` or ``qc.measure(q1, c1)`` will generate a synchronous
-    measurement pulse.
-
-    Args:
-        circuit: The quantum circuit to translate.
-        schedule_config: Backend specific parameters used for building the Schedule.
-
-    Returns:
-        A list of CircuitPulseDefs: the pulse definition for each circuit element.
-
-    Raises:
-        QiskitError: If circuit uses a command that isn't defined in config.inst_map.
-    """
-    circ_pulse_defs = []
-
-    inst_map = schedule_config.inst_map
-    qubit_mem_slots = {}  # Map measured qubit index to classical bit index
-
-    def get_measure_schedule() -> CircuitPulseDef:
-        """Create a schedule to measure the qubits queued for measuring."""
-        sched = Schedule()
-        sched += measure(qubits=list(qubit_mem_slots.keys()),
-                         inst_map=inst_map,
-                         meas_map=schedule_config.meas_map,
-                         qubit_mem_slots=qubit_mem_slots)
-        qubit_mem_slots.clear()
-        return CircuitPulseDef(schedule=sched,
-                               qubits=[chan.index for chan in sched.channels
-                                       if isinstance(chan, AcquireChannel)])
-
-    for inst, qubits, clbits in circuit.data:
-        inst_qubits = [qubit.index for qubit in qubits]  # We want only the indices of the qubits
-        if any(q in qubit_mem_slots for q in inst_qubits):
-            # If we are operating on a qubit that was scheduled to be measured, process that first
-            circ_pulse_defs.append(get_measure_schedule())
-        if isinstance(inst, Barrier):
-            circ_pulse_defs.append(CircuitPulseDef(schedule=inst, qubits=inst_qubits))
-        elif isinstance(inst, Measure):
-            if (len(inst_qubits) != 1 and len(clbits) != 1):
-                raise QiskitError("Qubit '{}' or classical bit '{}' errored because the "
-                                  "circuit Measure instruction only takes one of "
-                                  "each.".format(inst_qubits, clbits))
-            qubit_mem_slots[inst_qubits[0]] = clbits[0].index
-        else:
-            try:
-                circ_pulse_defs.append(
-                    CircuitPulseDef(schedule=inst_map.get(inst.name, inst_qubits, *inst.params),
-                                    qubits=inst_qubits))
-            except PulseError:
-                raise QiskitError("Operation '{}' on qubit(s) {} not supported by the backend "
-                                  "command definition. Did you remember to transpile your input "
-                                  "circuit for the same backend?".format(inst.name, inst_qubits))
-    if qubit_mem_slots:
-        circ_pulse_defs.append(get_measure_schedule())
-
-    return circ_pulse_defs
